requirements.txt
google-generativeai
pillow
python-dotenv
flask
